+++ packages/engine/src/index.ts
@@ -53,6 +53,7 @@ export * from "./rng";
 export * from "./blend";
 export * from "./nodeMapping";
 export * from "./layerStack";
+export * from "./layerSettings";
 export * from "./renderFrame";
 export * from "./valueCurve";
 export * from "./transition";